Refactor user specialties to user filters and search functionality
 • Replace UserControllerSpecialties with UserControllerFilters for user filtering by profession
 • Add UserControllerSearch for user search functionality
 • Remove old filter services and tests for days, locations, products, and specialties
 • Consolidate filter logic into UserServiceFilters
 • Add tests for new filter and search services
